Let GC use PATCH to delete finalizer from objects
