+++ pkg/controller/garbagecollector/operations.go
@@ -17,17 +17,12 @@ limitations under the License.
 package garbagecollector
 
 import (
-	"fmt"
-
 	"github.com/golang/glog"
 
-	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/util/retry"
 )
 
 // apiResource consults the REST mapper to translate an <apiVersion, kind,
@@ -100,44 +95,4 @@ func (gc *GarbageCollector) patchObject(item objectReference, patch []byte) (*un
 		return nil, err
 	}
 	return client.Resource(resource, item.Namespace).Patch(item.Name, types.StrategicMergePatchType, patch)
-}
-
-// TODO: Using Patch when strategicmerge supports deleting an entry from a
-// slice of a base type.
-func (gc *GarbageCollector) removeFinalizer(owner *node, targetFinalizer string) error {
-	err := retry.RetryOnConflict(retry.DefaultBackoff, func() error {
-		ownerObject, err := gc.getObject(owner.identity)
-		if errors.IsNotFound(err) {
-			return nil
-		}
-		if err != nil {
-			return fmt.Errorf("cannot finalize owner %s, because cannot get it: %v. The garbage collector will retry later.", owner.identity, err)
-		}
-		accessor, err := meta.Accessor(ownerObject)
-		if err != nil {
-			return fmt.Errorf("cannot access the owner object %v: %v. The garbage collector will retry later.", ownerObject, err)
-		}
-		finalizers := accessor.GetFinalizers()
-		var newFinalizers []string
-		found := false
-		for _, f := range finalizers {
-			if f == targetFinalizer {
-				found = true
-				continue
-			}
-			newFinalizers = append(newFinalizers, f)
-		}
-		if !found {
-			glog.V(5).Infof("the orphan finalizer is already removed from object %s", owner.identity)
-			return nil
-		}
-		// remove the owner from dependent's OwnerReferences
-		ownerObject.SetFinalizers(newFinalizers)
-		_, err = gc.updateObject(owner.identity, ownerObject)
-		return err
-	})
-	if errors.IsConflict(err) {
-		return fmt.Errorf("updateMaxRetries(%d) has reached. The garbage collector will retry later for owner %v.", retry.DefaultBackoff.Steps, owner.identity)
-	}
-	return err
 }
